Made some small changes to the login page, fixed padding on mobile
Cleaned up the login page's css a bit and added in a background behind the login panel. The background color looks interesting but might have to change a bit.

I looked for the reason that the login panel was not taking up the entire with of mobile devices. This apparently had to do with padding and this is now set to 0.

There was an issue where on some widths of the page the word 'Roosterwijzigingen' on the login page would not fit inside the login box. I fixed this by making the login panel a bit (1 column) wider.