Added CSS styling to the changes.php page
Worked on CSS styling for the changes.php page. A shown item is now color coded, a general item would have a green line before it while a warning would have an orange color before it.

Also added in a top-bar where it says 'Roosterwijzigingen' and it gives you the option to change from today to tomorrow or the other way around. It disabled the button you are on now and uses the post method to store the day of the week. This so that we avoid having to set the 'leerlingId' as well when we submit the form to go to another page.